Separated systems into their own module for platformer example
